커밋할게 너무쌓여서 그냥 할건데 사실 한게없음;; 	modified:   README.md 	modified:   src/main/java/com/example/ziindrive/file/controller/FileApiController.java 	modified:   src/main/java/com/example/ziindrive/file/controller/FileViewController.java 	modified:   src/main/java/com/example/ziindrive/user/controller/UserApiController.java 	modified:   src/main/resources/static/index.html 	modified:   src/main/resources/static/js/binRender.js 	modified:   src/main/resources/static/js/login.js 	modified:   src/main/resources/static/js/render.js 	renamed:    src/main/resources/templates/error.html -> src/main/resources/templates/error/4xx.html 	new file:   src/main/resources/templates/error/5xx.html 	modified:   src/main/resources/templates/login/loginView.html

+++ src/main/java/com/example/ziindrive/file/controller/FileApiController.java
@@ -39,6 +39,10 @@ public class FileApiController {
 
         // test
         System.out.println("* 아이디: " + userId);
+
+        if (userId == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok().body(service.findWithOptions(userId));
     }
 
@@ -144,6 +148,11 @@ public class FileApiController {
     public ResponseEntity<List<FileResponseDto>> deleteFile(
             @PathVariable("userId") String userId,
             @PathVariable("id") Long id) {
+
+        if (userId == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
         try {
             service.deleteFile(id);
             // 삭제한거 빼고 다시 검색한 결과 보내기
